Renamed AudioVisualizer1 file to Heathens and started work on circle that moves with the beat
