+++ java/src/ie/tudublin/Main.java
@@ -1,6 +1,7 @@
 package ie.tudublin;
 
 import C21423244.Monitor;
+import C21318233.Heathens;
 import example.CubeVisual;
 import example.Test;
 
@@ -13,6 +14,13 @@ public class Main
         processing.core.PApplet.runSketch( a, new Monitor());		
 	}
 
+	// For testing Heathens visualizer
+	public void heathens()
+	{
+		String[] a = {"MAIN"};
+        processing.core.PApplet.runSketch( a, new Heathens());		
+	}
+
 	public void testUI()
 	{
 		String[] a = {"MAIN"};
@@ -22,6 +30,6 @@ public class Main
 	public static void main(String[] args)
 	{
 		Main main = new Main();
-		main.startUI();			
+		main.startUI();
 	}
 }
